lead_collector/src/requirements.txt
playwright
python-dotenv
sqlalchemy
psycopg2-binary